Refactor enums in cluster api
