+++ cluster-api/src/main/java/io/aiven/klaw/clusterapi/models/ClusterAclRequest.java
@@ -2,6 +2,9 @@ package io.aiven.klaw.clusterapi.models;
 
 import java.io.Serializable;
 import javax.validation.constraints.NotNull;
+
+import io.aiven.klaw.clusterapi.models.enums.KafkaSupportedProtocol;
+import io.aiven.klaw.clusterapi.models.enums.RequestOperationType;
 import lombok.Builder;
 import lombok.Getter;
