+++ cluster-api/src/main/java/io/aiven/klaw/clusterapi/models/ClusterConnectorRequest.java
@@ -2,6 +2,8 @@ package io.aiven.klaw.clusterapi.models;
 
 import java.io.Serializable;
 import javax.validation.constraints.NotNull;
+
+import io.aiven.klaw.clusterapi.models.enums.KafkaSupportedProtocol;
 import lombok.Builder;
 import lombok.Getter;
