+++ cluster-api/src/main/java/io/aiven/klaw/clusterapi/models/ClusterSchemaRequest.java
@@ -1,6 +1,8 @@
 package io.aiven.klaw.clusterapi.models;
 
 import java.io.Serializable;
+
+import io.aiven.klaw.clusterapi.models.enums.KafkaSupportedProtocol;
 import lombok.Builder;
 import lombok.Getter;
