+++ cluster-api/src/main/java/io/aiven/klaw/clusterapi/models/ClusterTopicRequest.java
@@ -2,6 +2,8 @@ package io.aiven.klaw.clusterapi.models;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.Map;
+
+import io.aiven.klaw.clusterapi.models.enums.KafkaSupportedProtocol;
 import lombok.Builder;
 import lombok.Getter;
