+++ cluster-api/src/main/java/io/aiven/klaw/clusterapi/models/enums/ClusterStatus.java
@@ -1,4 +1,4 @@
-package io.aiven.klaw.clusterapi.models;
+package io.aiven.klaw.clusterapi.models.enums;
 
 public enum ClusterStatus {
   OFFLINE("OFFLINE"),
